+++ src/__generated__/ArtworkQuery.graphql.ts
@@ -277,7 +277,13 @@ fragment CommercialButtons_artwork on Artwork {
   isAcquireable
   isOfferable
   isInquireable
+  isBiddable
+  sale {
+    isClosed
+    id
+  }
   ...BuyNowButton_artwork
+  ...BidButton_artwork
   ...MakeOfferButton_artwork
 }
 
@@ -314,6 +320,37 @@ fragment CommercialEditionSetInformation_artwork on Artwork {
 
 fragment BuyNowButton_artwork on Artwork {
   internalID
+}
+
+fragment BidButton_artwork on Artwork {
+  slug
+  sale {
+    id
+    registrationStatus {
+      qualifiedForBidding
+      id
+    }
+    isPreview
+    isOpen
+    isLiveOpen
+    isClosed
+    isRegistrationClosed
+  }
+  myLotStanding(live: true) {
+    mostRecentBid {
+      maxBid {
+        cents
+      }
+      id
+    }
+  }
+  saleArtwork {
+    increments {
+      cents
+      display
+    }
+    id
+  }
 }
 
 fragment MakeOfferButton_artwork on Artwork {
@@ -615,11 +652,18 @@ v20 = {
   "args": null,
   "storageKey": null
 },
-v21 = [
+v21 = {
+  "kind": "ScalarField",
+  "alias": null,
+  "name": "display",
+  "args": null,
+  "storageKey": null
+},
+v22 = [
   (v3/*: any*/),
   (v2/*: any*/)
 ],
-v22 = {
+v23 = {
   "kind": "LinkedField",
   "alias": null,
   "name": "dimensions",
@@ -644,24 +688,31 @@ v22 = {
     }
   ]
 },
-v23 = {
+v24 = {
   "kind": "ScalarField",
   "alias": null,
   "name": "isAcquireable",
   "args": null,
   "storageKey": null
 },
-v24 = {
+v25 = {
   "kind": "ScalarField",
   "alias": null,
   "name": "isOfferable",
   "args": null,
   "storageKey": null
 },
-v25 = {
+v26 = {
   "kind": "ScalarField",
   "alias": null,
   "name": "saleMessage",
+  "args": null,
+  "storageKey": null
+},
+v27 = {
+  "kind": "ScalarField",
+  "alias": null,
+  "name": "cents",
   "args": null,
   "storageKey": null
 };
@@ -871,14 +922,61 @@ return {
               {
                 "kind": "ScalarField",
                 "alias": null,
-                "name": "isAuction",
+                "name": "isClosed",
+                "args": null,
+                "storageKey": null
+              },
+              {
+                "kind": "LinkedField",
+                "alias": null,
+                "name": "registrationStatus",
+                "storageKey": null,
+                "args": null,
+                "concreteType": "Bidder",
+                "plural": false,
+                "selections": [
+                  {
+                    "kind": "ScalarField",
+                    "alias": null,
+                    "name": "qualifiedForBidding",
+                    "args": null,
+                    "storageKey": null
+                  },
+                  (v2/*: any*/)
+                ]
+              },
+              {
+                "kind": "ScalarField",
+                "alias": null,
+                "name": "isPreview",
                 "args": null,
                 "storageKey": null
               },
               {
                 "kind": "ScalarField",
                 "alias": null,
-                "name": "isClosed",
+                "name": "isOpen",
+                "args": null,
+                "storageKey": null
+              },
+              {
+                "kind": "ScalarField",
+                "alias": null,
+                "name": "isLiveOpen",
+                "args": null,
+                "storageKey": null
+              },
+              {
+                "kind": "ScalarField",
+                "alias": null,
+                "name": "isRegistrationClosed",
+                "args": null,
+                "storageKey": null
+              },
+              {
+                "kind": "ScalarField",
+                "alias": null,
+                "name": "isAuction",
                 "args": null,
                 "storageKey": null
               }
@@ -1164,13 +1262,7 @@ return {
                                 "concreteType": "SaleArtworkCurrentBid",
                                 "plural": false,
                                 "selections": [
-                                  {
-                                    "kind": "ScalarField",
-                                    "alias": null,
-                                    "name": "display",
-                                    "args": null,
-                                    "storageKey": null
-                                  }
+                                  (v21/*: any*/)
                                 ]
                               },
                               (v2/*: any*/)
@@ -1190,7 +1282,7 @@ return {
                             ],
                             "concreteType": "Artist",
                             "plural": true,
-                            "selections": (v21/*: any*/)
+                            "selections": (v22/*: any*/)
                           },
                           {
                             "kind": "LinkedField",
@@ -1200,7 +1292,7 @@ return {
                             "args": null,
                             "concreteType": "Partner",
                             "plural": false,
-                            "selections": (v21/*: any*/)
+                            "selections": (v22/*: any*/)
                           },
                           (v6/*: any*/)
                         ]
@@ -1366,7 +1458,7 @@ return {
             "args": null,
             "storageKey": null
           },
-          (v22/*: any*/),
+          (v23/*: any*/),
           {
             "kind": "ScalarField",
             "alias": "edition_of",
@@ -1434,9 +1526,9 @@ return {
             "concreteType": "EditionSet",
             "plural": true,
             "selections": [
-              (v23/*: any*/),
               (v24/*: any*/),
               (v25/*: any*/),
+              (v26/*: any*/),
               (v2/*: any*/),
               (v5/*: any*/),
               {
@@ -1446,10 +1538,10 @@ return {
                 "args": null,
                 "storageKey": null
               },
-              (v22/*: any*/)
+              (v23/*: any*/)
             ]
           },
-          (v25/*: any*/),
+          (v26/*: any*/),
           {
             "kind": "ScalarField",
             "alias": null,
@@ -1464,14 +1556,87 @@ return {
             "args": null,
             "storageKey": null
           },
-          (v23/*: any*/),
           (v24/*: any*/),
+          (v25/*: any*/),
           {
             "kind": "ScalarField",
             "alias": null,
             "name": "isInquireable",
             "args": null,
             "storageKey": null
+          },
+          {
+            "kind": "ScalarField",
+            "alias": null,
+            "name": "isBiddable",
+            "args": null,
+            "storageKey": null
+          },
+          {
+            "kind": "LinkedField",
+            "alias": null,
+            "name": "myLotStanding",
+            "storageKey": "myLotStanding(live:true)",
+            "args": [
+              {
+                "kind": "Literal",
+                "name": "live",
+                "value": true
+              }
+            ],
+            "concreteType": "LotStanding",
+            "plural": true,
+            "selections": [
+              {
+                "kind": "LinkedField",
+                "alias": null,
+                "name": "mostRecentBid",
+                "storageKey": null,
+                "args": null,
+                "concreteType": "BidderPosition",
+                "plural": false,
+                "selections": [
+                  {
+                    "kind": "LinkedField",
+                    "alias": null,
+                    "name": "maxBid",
+                    "storageKey": null,
+                    "args": null,
+                    "concreteType": "BidderPositionMaxBid",
+                    "plural": false,
+                    "selections": [
+                      (v27/*: any*/)
+                    ]
+                  },
+                  (v2/*: any*/)
+                ]
+              }
+            ]
+          },
+          {
+            "kind": "LinkedField",
+            "alias": null,
+            "name": "saleArtwork",
+            "storageKey": null,
+            "args": null,
+            "concreteType": "SaleArtwork",
+            "plural": false,
+            "selections": [
+              {
+                "kind": "LinkedField",
+                "alias": null,
+                "name": "increments",
+                "storageKey": null,
+                "args": null,
+                "concreteType": "BidIncrementsFormatted",
+                "plural": true,
+                "selections": [
+                  (v27/*: any*/),
+                  (v21/*: any*/)
+                ]
+              },
+              (v2/*: any*/)
+            ]
           }
         ]
       }
@@ -1480,7 +1645,7 @@ return {
   "params": {
     "operationKind": "query",
     "name": "ArtworkQuery",
-    "id": "1d1b52edb24e07dbe1e74a7be1c5f792",
+    "id": "5877a62793253ff61d1b25e8ef97991e",
     "text": null,
     "metadata": {}
   }